+++ lib/remote_ruby/source_extractor.rb
@@ -1,24 +1,14 @@
 # frozen_string_literal: true
 
 require 'method_source'
-require 'parser/current'
-require 'unparser'
-
-# Opt-in to most recent AST format
-Parser::Builders::Default.emit_lambda              = true
-Parser::Builders::Default.emit_procarg0            = true
-Parser::Builders::Default.emit_encoding            = true
-Parser::Builders::Default.emit_index               = true
-Parser::Builders::Default.emit_arg_inside_procarg0 = true
-Parser::Builders::Default.emit_forward_arg         = true
-Parser::Builders::Default.emit_kwargs              = true
-Parser::Builders::Default.emit_match_pattern       = true
+require 'remote_ruby/parser_factory'
 
 module RemoteRuby
   # Receives a block and extracts Ruby code (as a string) with this block's
   # source
   class SourceExtractor
     def extract(&block)
+      RemoteRuby::ParserFactory.require_parser
       ast = Parser::CurrentRuby.parse(block.source)
       block_node = find_block(ast)